Generate TAGS with relative filenames when appropriate
The TAGS file can list filenames with either absolute or relative
filenames. If the filenames are relative, they are considered relative
in relation to the TAGS file itself. This allows a TAGS file that is
located inside a project's basedir to be moved with the project
itself.

The change detects if the TAGS file is in the basedir, and if so,
generates relative filenames in the TAGS file. It also takes care to
set default-directory properly: the find|etags command should run from
the directory containing the TAGS file.