project-find-file{-ido} presents relative path names
The path names presented by project-find-file (and the -ido version)
are now relative to the project's basedir. This cuts down on visual
clutter and makes it easier to select the file of interest.

project-find-file (and -ido) do not care if the file-index contains
absolute or relative path names as the job of finding matches in
file-index is delegated to a new utility function used by both.

Additionally, the file-file functions now simply warn if the
file-index buffer doesn't exist (and then return) rather than offering
to generate the index. The former method was confusing because even
though the index was generated at the user's request, the find-file
operation had to abort as project-index runs in the background.